The Visual C++ version now works properly (there were some linking problems due to that silly compiler).
git-svn-id: file:///mnt/stora_backup/satt/backup-before-new-install/var/svn/repos/software/EasyLocalpp/tags/2.0-pre@188 1ceef749-5ca8-4b7e-9fa7-ff5341897c31

+++ src/helpers/StateManager.hh
@@ -55,7 +55,7 @@ public:
   virtual bool CheckConsistency(const State& st) const = 0;
   
   CostComponent<Input, State,CFtype>& GetCostComponent(unsigned i) const { return *(cost_component[i]); }
-  unsigned CostComponents() const { return cost_component.size(); }
+  size_t CostComponents() const { return cost_component.size(); }
   CFtype Cost(const State& st, unsigned int i) const { return cost_component[i]->Cost(st); }
 
   const std::string name;
